+++ secp256k1-sys/src/lib.rs
@@ -133,6 +133,7 @@ impl SchnorrSigExtraParams {
 /// Library-internal representation of a Secp256k1 public key
 #[repr(C)]
 #[derive(Copy, Clone)]
+#[cfg_attr(fuzzing, derive(PartialEq, Eq, PartialOrd, Ord, Hash))]
 pub struct PublicKey([c_uchar; 64]);
 impl_array_newtype!(PublicKey, c_uchar, 64);
 impl_raw_debug!(PublicKey);
@@ -227,6 +228,7 @@ impl core::hash::Hash for PublicKey {
 /// Library-internal representation of a Secp256k1 signature
 #[repr(C)]
 #[derive(Copy, Clone)]
+#[cfg_attr(fuzzing, derive(PartialEq, Eq, PartialOrd, Ord, Hash))]
 pub struct Signature([c_uchar; 64]);
 impl_array_newtype!(Signature, c_uchar, 64);
 impl_raw_debug!(Signature);
@@ -315,6 +317,7 @@ impl core::hash::Hash for Signature {
 
 #[repr(C)]
 #[derive(Copy, Clone)]
+#[cfg_attr(fuzzing, derive(PartialEq, Eq, PartialOrd, Ord, Hash))]
 pub struct XOnlyPublicKey([c_uchar; 64]);
 impl_array_newtype!(XOnlyPublicKey, c_uchar, 64);
 impl_raw_debug!(XOnlyPublicKey);
@@ -404,6 +407,7 @@ impl core::hash::Hash for XOnlyPublicKey {
 
 #[repr(C)]
 #[derive(Copy, Clone)]
+#[cfg_attr(fuzzing, derive(PartialEq, Eq, PartialOrd, Ord, Hash))]
 pub struct KeyPair([c_uchar; 96]);
 impl_array_newtype!(KeyPair, c_uchar, 96);
 impl_raw_debug!(KeyPair);